# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/orders": {
      "post": {
        "operationId": "OrdersController_createOrder",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateOrderResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body."
          },
          "401": {
            "description": "Unauthorized."
          },
          "404": {
            "description": "Restaurant not found."
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "summary": "Create a new order for the authenticated user",
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/orders/{id}": {
      "get": {
        "operationId": "OrdersController_findOrderById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the order details.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDetailsDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid UUID format."
          },
          "401": {
            "description": "Unauthorized."
          },
          "404": {
            "description": "Order not found."
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "summary": "Get details of a specific order belonging to the authenticated user",
        "tags": [
          "Orders"
        ]
      }
    }
  },
  "info": {
    "title": "MFOS Backend API",
    "description": "Multi-tenant Food Ordering System backend API.\n\nAll protected endpoints require a Bearer JWT token.\nObtain a token via `POST /api/auth/login` or `POST /api/auth/register`, then click **Authorize** and enter `Bearer <token>`.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "jwt": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateOrderItemDto": {
        "type": "object",
        "properties": {
          "menuItemId": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000",
            "description": "UUID of the menu item"
          },
          "quantity": {
            "type": "number",
            "example": 2,
            "description": "Quantity to order, minimum 1"
          }
        },
        "required": [
          "menuItemId",
          "quantity"
        ]
      },
      "CreateOrderDto": {
        "type": "object",
        "properties": {
          "restaurantId": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440001",
            "description": "UUID of the restaurant"
          },
          "items": {
            "description": "Non-empty list of items to order",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateOrderItemDto"
            }
          }
        },
        "required": [
          "restaurantId",
          "items"
        ]
      },
      "CreateOrderResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440002",
            "description": "UUID of the created order"
          },
          "status": {
            "type": "string",
            "enum": [
              "CREATED",
              "PENDING"
            ],
            "example": "PENDING",
            "description": "Current status of the order"
          },
          "totalPrice": {
            "type": "number",
            "example": 5980,
            "description": "Total price in HUF"
          }
        },
        "required": [
          "id",
          "status",
          "totalPrice"
        ]
      },
      "OrderRestaurantDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440001",
            "description": "UUID of the restaurant"
          },
          "name": {
            "type": "string",
            "example": "Burger House",
            "description": "Name of the restaurant"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "OrderItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440003",
            "description": "UUID of the order item"
          },
          "name": {
            "type": "string",
            "example": "Classic Burger",
            "description": "Name of the menu item at time of order"
          },
          "price": {
            "type": "number",
            "example": 2990,
            "description": "Price of the menu item in HUF at time of order"
          },
          "quantity": {
            "type": "number",
            "example": 2,
            "description": "Quantity ordered"
          }
        },
        "required": [
          "id",
          "name",
          "price",
          "quantity"
        ]
      },
      "OrderDetailsDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440002",
            "description": "UUID of the order"
          },
          "status": {
            "type": "string",
            "enum": [
              "CREATED",
              "PENDING"
            ],
            "example": "PENDING",
            "description": "Current status of the order"
          },
          "totalPrice": {
            "type": "number",
            "example": 5980,
            "description": "Total price in HUF"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-05-04T13:40:29.227Z",
            "description": "Timestamp when the order was created"
          },
          "restaurant": {
            "description": "Restaurant the order was placed at",
            "allOf": [
              {
                "$ref": "#/components/schemas/OrderRestaurantDto"
              }
            ]
          },
          "items": {
            "description": "Items included in the order",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItemDto"
            }
          }
        },
        "required": [
          "id",
          "status",
          "totalPrice",
          "createdAt",
          "restaurant",
          "items"
        ]
      }
    }
  }
}